Split network module into a separate file.
